Rama terminada

+++ src/main/java/com/salesianostriana/dam/trianafy/service/ArtistService.java
@@ -1,6 +1,8 @@
 package com.salesianostriana.dam.trianafy.service;
 
 
+import com.salesianostriana.dam.trianafy.exception.ArtistNotFoundException;
+import com.salesianostriana.dam.trianafy.exception.EmptyArtistListException;
 import com.salesianostriana.dam.trianafy.model.Artist;
 import com.salesianostriana.dam.trianafy.repos.ArtistRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,21 +17,30 @@ public class ArtistService {
 
     private final ArtistRepository repository;
 
+    private final SongService songService;
+
 
     public Artist add(Artist artist) {
         return repository.save(artist);
     }
 
-    public Optional<Artist> findById(Long id) {
-        return repository.findById(id);
+    public Artist findById(Long id) {
+        return repository.findById(id).orElseThrow(() -> new ArtistNotFoundException(id));
     }
 
     public List<Artist> findAll() {
-        return repository.findAll();
+        List<Artist> result = repository.findAll();
+        if(result.isEmpty()) {
+            throw new EmptyArtistListException();
+        }
+        return result;
     }
 
-    public Artist edit(Artist artist) {
-        return repository.save(artist);
+    public Artist edit(Long id, Artist edited) {
+        return repository.findById(id).map(artist -> {
+            artist.setName(edited.getName());
+            return repository.save(artist);
+        }).orElseThrow(() -> new ArtistNotFoundException(id));
     }
 
     public void delete(Artist artist) {
@@ -37,7 +48,13 @@ public class ArtistService {
     }
 
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        if(repository.existsById(id)) {
+            songService.findByArtist(findById(id))
+                    .stream()
+                            .forEach(song -> song.setArtist(null));
+            repository.deleteById(id);
+
+        }
     }
 
     public boolean existsById(Long id) {
